+++ nbproject/project.properties
@@ -76,7 +76,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=FamilyBot
-main.class=familybot.App
+main.class=cu.edu.cujae.ceis.familibot.App
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
