+++ src/cu/edu/cujae/ceis/familybot/logic/controllers/SimulationController.java
@@ -1,7 +1,7 @@
-package familybot.logic.controllers;
+package cu.edu.cujae.ceis.familybot.logic.controllers;
 
-import familybot.logic.core.Board;
-import familybot.logic.core.Family;
+import cu.edu.cujae.ceis.familybot.logic.core.Board;
+import cu.edu.cujae.ceis.familybot.logic.core.Family;
 import java.util.ArrayList;
 
 public class SimulationController
