+++ src/cu/edu/cujae/ceis/familybot/logic/core/Board.java
@@ -2,9 +2,9 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package familybot.logic.core;
+package cu.edu.cujae.ceis.familybot.logic.core;
 
-import familybot.logic.exceptions.NotInitializedException;
+import cu.edu.cujae.ceis.familybot.logic.exceptions.NotInitializedException;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
