+++ src/cu/edu/cujae/ceis/familybot/logic/core/Direction.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Enum.java to edit this template
  */
-package familybot.logic.core;
+package cu.edu.cujae.ceis.familybot.logic.core;
 
 /**
  *
