+++ src/cu/edu/cujae/ceis/familybot/logic/exceptions/NotInitializedException.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package familybot.logic.exceptions;
+package cu.edu.cujae.ceis.familybot.logic.exceptions;
 
 /**
  *
