+++ src/cu/edu/cujae/ceis/familybot/logic/utils/CharGenerator.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package familybot.logic.utils;
+package cu.edu.cujae.ceis.familybot.logic.utils;
 
 import java.io.Serializable;
