+++ src/cu/edu/cujae/ceis/familybot/logic/utils/SimulationRecord.java
@@ -1,7 +1,7 @@
-package familybot.logic.utils;
+package cu.edu.cujae.ceis.familybot.logic.utils;
 
-import familybot.logic.core.Direction;
-import familybot.logic.core.Robot;
+import cu.edu.cujae.ceis.familybot.logic.core.Direction;
+import cu.edu.cujae.ceis.familybot.logic.core.Robot;
 import java.io.Serializable;
 import java.util.HashMap;
 import java.util.List;
